// src/types/user/address.ts
export interface IAddress {
  id?: string;
  user?: string;
  phone?: string;
  fullName?: string;
  city?: string;
  address?: string;
  subAddress?: string;
  label?: string;
  isDefault?: string;
  status?: string;
  isDeleted?: boolean;
  deleteAt?: string;
  dateCreated?: string;
  dateUpdate?: string;
}